Journal latency type should be double
Former-commit-id: 754e715e7504cd51aeaac432e31cb5e2870458a1

+++ src/main/java/io/specto/hoverfly/junit/core/model/JournalEntry.java
@@ -20,14 +20,14 @@ public class JournalEntry {
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSXXX")
     private final LocalDateTime timeStarted;
 
-    private final Integer latency;
+    private final Double latency;
 
     @JsonCreator
     public JournalEntry(@JsonProperty("request") RequestDetails request,
                         @JsonProperty("response") Response response,
                         @JsonProperty("mode") String mode,
                         @JsonProperty("timeStarted") LocalDateTime timeStarted,
-                        @JsonProperty("latency") Integer latency) {
+                        @JsonProperty("latency") Double latency) {
         this.request = request;
         this.response = response;
         this.mode = mode;
@@ -52,7 +52,7 @@ public class JournalEntry {
         return timeStarted;
     }
 
-    public Integer getLatency() {
+    public Double getLatency() {
         return latency;
     }
 }
